00. AI 놀이터/exam-digitalizer/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exam-digitalizer-backend"
version = "0.1.0"
description = "시험 문항 디지털라이징 시스템 백엔드"
requires-python = ">=3.12"

dependencies = [
    # Web Framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",

    # DB
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pgvector>=0.2.5",
    "psycopg2-binary>=2.9.9",

    # Redis
    "redis[hiredis]>=5.0.4",

    # S3 / MinIO
    "boto3>=1.34.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",

    # Config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "email-validator>=2.1.0",

    # LLM
    "anthropic>=0.26.0",
    "openai>=1.30.0",

    # HWP
    "lxml>=5.2.1",
    "Pillow>=10.3.0",

    # Utils
    "structlog>=24.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "agents*", "core*", "models*", "schemas*", "migrations*", "scripts*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
